EEV.c: - fixed/finalized stepper - added some debug info

+++ analog.h
@@ -12,7 +12,7 @@
 
 #include <stdint.h>
 
-extern uint16_t p;
+extern uint32_t p;
 
 void analog_init(void);
